refactor(signIn): load internal components asynchronously
